third_party/blink/web_tests/storage/domstorage/events/case-sensitive-expected.txt

Verify that storage events fire even when only the case of the value changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing sessionStorage
storage.clear()
PASS storage.length is 0


Verify storage events are case sensitive
storage.foo = 'test'
Reset storage event list
storageEventList = new Array()
storage.foo = 'test'
PASS storageEventList.length is 0
storage.foo = 'TEST'
PASS storageEventList.length is 1




Testing localStorage
storage.clear()
PASS storage.length is 0


Verify storage events are case sensitive
storage.foo = 'test'
Reset storage event list
storageEventList = new Array()
storage.foo = 'test'
PASS storageEventList.length is 0
storage.foo = 'TEST'
PASS storageEventList.length is 1
PASS successfullyParsed is true

TEST COMPLETE
